requirements-worker.txt
# Standalone Worker Requirements
# =============================
# Minimal dependencies for running gRINN workers remotely

# Core Celery and task processing
celery==5.3.4
redis==5.0.1

# Database connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Docker integration
docker==6.1.3

# Logging and utilities
python-dotenv==1.0.0
requests==2.31.0

# Development dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0